Añadida macro para habilitar o desabilitar excepciones dentro del código

+++ lib/AnalogOutput/AnalogOutput.cpp
@@ -4,6 +4,6 @@ AnalogOutput::AnalogOutput(uint8_t pin)
     : Pin{ pin, OUTPUT } {
 }
 
-void AnalogOutput::write(uint16_t value) {
+void AnalogOutput::write(int value) {
     analogWrite(this->getPin(), value);
 }
